refactor: extract error formatting logic to separate errors.ts file
